requirements.txt
anyascii==0.3.2
beautifulsoup4==4.12.3
blinker==1.9.0
cbor==1.0.0
certifi==2024.12.14
charset-normalizer==3.4.0
chest==0.2.3
click==8.1.7
colorama==0.4.6
contractions==0.1.73
Deprecated==1.2.15
dill==0.3.9
Flask==3.1.0
gensim==4.3.3
gunicorn==23.0.0
HeapDict==1.0.1
idna==3.10
ijson==3.3.0
inscriptis==2.5.0
ir_datasets==0.5.9
ir_measures==0.3.6
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
lxml==5.3.0
lz4==4.3.3
MarkupSafe==3.0.2
more-itertools==10.5.0
numpy==1.26.4
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pyahocorasick==2.1.0
pyjnius==1.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-terrier==0.12.1
pytrec-eval-terrier==0.5.6
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.6.0
scipy==1.13.1
six==1.17.0
smart-open==7.1.0
soupsieve==2.6
statsmodels==0.14.4
textsearch==0.0.24
threadpoolctl==3.5.0
tqdm==4.67.1
trec-car-tools==2.6
tzdata==2024.2
unlzw3==0.2.3
urllib3==2.2.3
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
Werkzeug==3.1.3
wget==3.2
wrapt==1.17.0
xgboost==2.1.3
zlib-state==0.1.9
